# Make createTimeWithContext faster on MacOS