@PierreF 2022-01-12
I've just seen we already get the KprocInfo with `getKProc()` method on Process. I'll use `getKProc()`

Do you want me to pass though existing `getKProc()` (unmodified) or change `getKProc()` to use `SysctlKinfoProc` from 
golang.org/x/sys ?