@Lomanic 2022-01-12
Thanks a lot for this change, `unix.SysctlKinfoProc()`  was added to x/sys/unix in [`95da234e122c5f925d4f955e47843f027369c273`](https://go.googlesource.com/sys/+/95da234e122c5f925d4f955e47843f027369c273) two months ago.

> The KinfoProc may contains more (static) information like PPID which could be cached to avoid futher calls.

Ppid are not static on unix, orphan/zombie processes can get attached to pid 1 (init) if their parent is killed. [Process ppid is only cached and cacheable on windows](https://github.com/shirou/gopsutil/blob/1ccce7b91c1903b0e280547bb1b1a32aa815fc28/process/process_windows.go#L306).

> Do you want me to pass though existing getKProc() (unmodified) or change getKProc() to use SysctlKinfoProc from
golang.org/x/sys ?

Let's modify getKProc to use the `unix.SysctlKinfoProc()` function indeed.

I don't have any darwin host or VM these days so won't be able to test any of this.