@PierreF 2022-01-13
Indeed PPID can't be cached. Missed this one.

Superseeded this PR by #1226 which use KProc even for more thing that just createTime